chore: move ErrorCode.java
