+++ film-api/src/main/java/com/programmers/film/api/post/exception/MapLocationException.java
@@ -1,6 +1,6 @@
 package com.programmers.film.api.post.exception;
 
-import com.programmers.film.common.error.exception.ErrorCode;
+import com.programmers.film.common.error.ErrorCode;
 import com.programmers.film.common.error.exception.InvalidTypeException;
 
 public class MapLocationException extends InvalidTypeException {
