+++ film-api/src/main/java/com/programmers/film/api/post/exception/PostCanNotOpenException.java
@@ -1,7 +1,7 @@
 package com.programmers.film.api.post.exception;
 
+import com.programmers.film.common.error.ErrorCode;
 import com.programmers.film.common.error.exception.EntityNotFoundException;
-import com.programmers.film.common.error.exception.ErrorCode;
 
 public class PostCanNotOpenException extends EntityNotFoundException {
