+++ film-api/src/main/java/com/programmers/film/api/post/exception/PostIdNotFoundException.java
@@ -1,8 +1,7 @@
 package com.programmers.film.api.post.exception;
 
-import com.programmers.film.common.error.exception.BusinessException;
+import com.programmers.film.common.error.ErrorCode;
 import com.programmers.film.common.error.exception.EntityNotFoundException;
-import com.programmers.film.common.error.exception.ErrorCode;
 
 public class PostIdNotFoundException extends EntityNotFoundException {
     public PostIdNotFoundException(String message) {
