+++ film-api/src/main/java/com/programmers/film/api/post/exception/PostOpenableAtException.java
@@ -1,6 +1,6 @@
 package com.programmers.film.api.post.exception;
 
-import com.programmers.film.common.error.exception.ErrorCode;
+import com.programmers.film.common.error.ErrorCode;
 import com.programmers.film.common.error.exception.InvalidInputValueException;
 
 public class PostOpenableAtException extends InvalidInputValueException {
