+++ film-api/src/main/java/com/programmers/film/api/user/exception/UserIdNotFoundExceoption.java
@@ -1,7 +1,7 @@
 package com.programmers.film.api.user.exception;
 
+import com.programmers.film.common.error.ErrorCode;
 import com.programmers.film.common.error.exception.EntityNotFoundException;
-import com.programmers.film.common.error.exception.ErrorCode;
 
 public class UserIdNotFoundExceoption extends EntityNotFoundException {
     public UserIdNotFoundExceoption(String message) {
