+++ film-common/src/main/java/com/programmers/film/common/error/exception/BusinessException.java
@@ -1,5 +1,6 @@
 package com.programmers.film.common.error.exception;
 
+import com.programmers.film.common.error.ErrorCode;
 import lombok.Getter;
 
 @Getter
