+++ film-common/src/main/java/com/programmers/film/common/error/exception/EntityNotFoundException.java
@@ -1,13 +1,15 @@
 package com.programmers.film.common.error.exception;
 
+import com.programmers.film.common.error.ErrorCode;
+
 public class EntityNotFoundException extends BusinessException {
 
-    public EntityNotFoundException(String message, ErrorCode errorCode) {
-        super(message, errorCode);
-    }
+	public EntityNotFoundException(String message, ErrorCode errorCode) {
+		super(message, errorCode);
+	}
 
-    public EntityNotFoundException(String message) {
-        super(message, ErrorCode.ENTITY_NOT_FOUND);
-    }
+	public EntityNotFoundException(String message) {
+		super(message, ErrorCode.ENTITY_NOT_FOUND);
+	}
 }
